fix: program failing when unknown record type is in zone
The upstream `cloudflare` crate does not support parsing CAA records
(and others), which means that the result of the record listing endpoint
cannot be parsed by cfdyndns, resulting in a crash.

This PR adds a filter to the requests to only get A and AAAA records.
However, due to the design of the upstream crate, it is not possible to
filter by type without also filtering by content, meaning we have to
change the `match` parameter to `any`. Additionally, the API only allows
to filter by a single type, which means we duplicate the number of
requests by making one request for A records and another for AAAA
records.

Fixes #54